refactor(deps): split deps.py — extract context_types + dev_mode_setup (#92)
Resolves smell-scanner MEDIUM finding from PR #87 audit. ``deps.py``
had grown to **678 LOC** absorbing responsibility from #69, #70, and
#84 — type definitions, FastAPI Depends resolvers, dev-mode PAT
provisioning, oauth2-proxy header parsing, beta-access gating, scope
resolution, PAT-scope narrowing, and RLS GUC injection — with the
highest production-code consumer fan-in (~52 importers) of any
non-model file in the service. Any future auth refactor touched it
and rippled.

Two clean extractions, **zero caller-side churn** (re-exports preserve
the import surface):

## ``core/api/context_types.py`` (NEW, 65 LOC)

Owns the 4 frozen dataclasses:

- ``PatContext`` — PAT metadata
- ``CurrentUserContext`` — per-request user context (was the dict-replacement from #69)
- ``UserContext`` — pairs ORM user row with ``CurrentUserContext`` (from #84)
- ``RequestContext`` — adds ``ScopeContext`` on top (from #84)

Uses ``TYPE_CHECKING`` for the ``models.User`` and ``ScopeContext``
forward references so the file has no runtime DB dependency. ``deps.py``
re-exports them via ``from core.api.context_types import *`` so callers
can still ``from core.api.deps import UserContext`` unchanged.

## ``core/api/dev_mode_setup.py`` (NEW, 78 LOC)

Owns dev-mode user provisioning: superadmin rights + beta-access +
default PAT for ``dev@localhost``. Renamed the function from
``_ensure_dev_mode_defaults`` (private convention) to
``ensure_dev_mode_defaults`` (public, since it's now cross-module);
``deps.py`` imports it as ``_ensure_dev_mode_defaults`` for compat.

Also moves the ``_DEV_MODE_PAT_CACHE`` process-local cache and these
imports out of deps.py: ``token_repo``, ``parse_token``,
``TokenCreateRequest``.

## ``deps.py`` (678 → 596 LOC)

Top of file now reads cleanly: imports, dataclass re-exports,
``get_or_create_user_for_request``, then the resolvers. The 145 LOC
removed is the dataclass definitions + the dev-mode helper + their
supporting imports.

## Verification

- ``uv run pytest --cov=core --cov-fail-under=80``:
  - **875 passed**, 8 skipped (unchanged from baseline)
  - **Coverage 80.79%** ✓
- All 52 ``from core.api.deps import ...`` callers continue to work
  via re-exports — no caller-side migrations needed.

Closes #92.
